Fix starting a feature branch for a repo that already had changes for this branch
This is a rather 'out there' case, but the idempotent feature start command
was failing due to changes in the remote repo on the same feature branch that
was previously started.  You need to checkout and pull the latest into the
repo before attempting to push again.

This could also be fixed just pushing when starting the feature branch.